move the definition into a header file so it can be further derived
